feat: Add PrintAllItems
