--- App.tsx
@@ -1,12 +1,0 @@
-import { BrowserRouter } from "react-router-dom";
-import MarketPage from "./pages/MarketPage/MarketPage";
-
-function App() {
-  return (
-    <BrowserRouter>
-      <MarketPage />
-    </BrowserRouter>
-  );
-}
-
-export default App;
