--- pages/MarketPage/MarketPage.tsx
@@ -1,11 +1,0 @@
-import PrintBestItems from "./components/PrintBestItems";
-
-const MarketPage: React.FC = () => {
-  return (
-    <div>
-      <PrintBestItems />
-    </div>
-  );
-};
-
-export default MarketPage;
